+++ changelog.md
@@ -7,7 +7,8 @@
 3. Deduplicated roles and event attendees on update  
 4. Prevented disabling password login when there is no other login method  
 5. Added audit log for updating global note  
-6. Removed the need to login when using email confirmation link  
+6. Removed authorization requirement when using email confirmation link  
+7. Added German translation  
 
 ## v2.8.2